# static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "Wash Company SwaggerUi Api",
    "version": "1.0.0",
    "title": "WashCompany RestAPI"
  },
  "tags": [
    {
      "name": "User Auth",
      "description": "User authentication"
    },
    {
      "name": "Wash Company",
      "description": "Add and get company"
    },
    {
      "name": "Orders",
      "description": "Manipulate with orders"
    },
    {
      "name": "Washers",
      "description": "Manipulate with washers"
    }
  ],
  "servers": [
    {
      "url": "/api"
    }
  ],
  "components": {
    "schemas": {
      "signIn": {
        "type": "object",
        "properties": {
          "login": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "signUp": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "required": true
          },
          "login": {
            "type": "string",
            "required": true
          },
          "password": {
            "type": "string",
            "required": true
          },
          "role": {
            "type": "string",
            "required": true,
            "default": "admin",
            "enum": [
              "admin",
              "owner"
            ]
          }
        }
      },
      "addWashCompany": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "required": true
          },
          "avatar": {
            "type": "string",
            "description": "Company avatar in bytes format",
            "required": false
          },
          "location": {
            "type": "string",
            "required": true
          }
        }
      },
      "integerId": {
        "type": "integer",
        "properties": {
          "companyId": {
            "type": "integer"
          }
        }
      },
      "addOrder": {
        "type": "object",
        "properties": {
          "price": {
            "type": "integer",
            "required": true
          },
          "carModel": {
            "type": "string",
            "required": true
          },
          "carNumber": {
            "type": "string",
            "required": true
          },
          "clientName": {
            "type": "string",
            "required": true
          },
          "clientNumber": {
            "type": "integer",
            "required": true
          },
          "isActive": {
            "type": "boolean",
            "required": false,
            "default": true
          },
          "services": {
            "type": "array",
            "required": false,
            "default": []
          },
          "washers": {
            "type": "array",
            "required": false,
            "default": []
          }
        }
      },
      "editOrder": {
        "type": "object",
        "properties": {
          "price": {
            "type": "integer",
            "required": false
          },
          "carModel": {
            "type": "string",
            "required": false
          },
          "carNumber": {
            "type": "string",
            "required": false
          },
          "isActive": {
            "type": "boolean",
            "required": false
          },
          "isCancelled": {
            "type": "boolean",
            "required": false
          },
          "services": {
            "type": "array",
            "required": false
          },
          "washers": {
            "type": "array",
            "required": false
          }
        }
      },
      "addWasher": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "required": true
          },
          "phoneNumber": {
            "type": "integer",
            "required": true
          },
          "stake": {
            "type": "integer",
            "required": true
          },
          "isActive": {
            "type": "boolean",
            "default": true,
            "required": true
          }
        }
      },
      "editWasher": {"type": "object",
        "properties": {
          "name": {
            "type": "string",
            "required": false
          },
          "phoneNumber": {
            "type": "integer",
            "required": false
          },
          "stake": {
            "type": "integer",
            "required": false
          },
          "isActive": {
            "type": "boolean",
            "required": false,
            "default": true
          },
          "washCompanyId": {
            "type": "integer",
            "required": false
          },
          "orders": {
            "type": "array",
            "required": false,
            "default": []
          }
        }
      }
    }
  },
  "paths": {
    "/signIn": {
      "post": {
        "tags": [
          "User Auth"
        ],
        "summary": "Sign in for user",
        "requestBody": {
          "description": "Sign in by login and password",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/signIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authenticated"
          },
          "400": {
            "description": "Validation error"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/signUp": {
      "post": {
        "tags": [
          "User Auth"
        ],
        "summary": "Sign up for user",
        "requestBody": {
          "description": "Sign up by user data",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/signUp"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authenticated"
          },
          "400": {
            "description": "Validation error"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/signOut": {
      "post": {
        "tags": [
          "User Auth"
        ],
        "summary": "Sign out for user",
        "requestBody": {
          "description": "Sign out without data",
          "required": false,
          "content": {}
        },
        "responses": {
          "200": {
            "description": "Authenticated"
          },
          "400": {
            "description": "Validation error"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/washCompanyInsert": {
      "post": {
        "tags": [
          "Wash Company"
        ],
        "summary": "Add new Wash Company",
        "requestBody": {
          "description": "Admin or owner can add new companies",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/addWashCompany"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authenticated"
          },
          "400": {
            "description": "Validation error"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/getWashCompanyId": {
      "post": {
        "tags": [
          "Wash Company"
        ],
        "summary": "Get Wash Company id",
        "requestBody": {
          "description": "Get users WashCompany id",
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/signIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authenticated"
          },
          "400": {
            "description": "Validation error"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/{WashCompanyId}/orders": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "Get company orders",
        "parameters": [
          {
            "name": "WashCompanyId",
            "in": "path",
            "description": "WashCompanyId for get orders",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/integerId"
            }
          },
          {
            "name": "IsActive",
            "in": "query",
            "required": true,
            "description": "Orders status",
            "schema": {
              "type": "boolean",
              "properties": {
                "IsActive": {
                  "type": "boolean",
                  "default": true
                }
              }
            }
          },
          {
            "name": "dateFrom",
            "in": "query",
            "required": true,
            "description": "Orders from date (year-month-day)",
            "schema": {
              "type": "string",
              "properties": {
                "dateFrom": {
                  "type": "string",
                  "default": "year-month-day",
                  "required": true
                }
              }
            }
          },
          {
            "name": "dateTo",
            "in": "query",
            "required": true,
            "description": "Orders to date (year-month-day)",
            "schema": {
              "type": "string",
              "properties": {
                "dateTo": {
                  "type": "string",
                  "default": "year-month-day",
                  "required": true
                }
              }
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": true,
            "description": "Pagination page number (start from 1)",
            "schema": {
              "type": "integer",
              "properties": {
                "page": {
                  "type": "integer",
                  "default": 1
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authenticated"
          },
          "400": {
            "description": "Validation error"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/order/{orderId}": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "Get order info",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "description": "Order id for getting info",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/integerId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authenticated"
          },
          "400": {
            "description": "Validation error"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/{WashCompanyId}/insertOrder": {
      "post": {
        "tags": [
          "Orders"
        ],
        "summary": "Add new order for company",
        "parameters": [
          {
            "name": "WashCompanyId",
            "in": "path",
            "description": "CompanyId for adding new order",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/integerId"
            }
          }
        ],
        "requestBody": {
          "description": "Add new order for this company",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/addOrder"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authenticated"
          },
          "400": {
            "description": "Validation error"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/{orderId}/updateOrder": {
      "post": {
        "tags": [
          "Orders"
        ],
        "summary": "Update exists order",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "description": "OrderId for editing order",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/integerId"
            }
          }
        ],
        "requestBody": {
          "description": "Edit order",
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/editOrder"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authenticated"
          },
          "400": {
            "description": "Validation error"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/{WashCompanyId}/washers": {
      "get": {
        "tags": [
          "Washers"
        ],
        "summary": "Get washers by name",
        "parameters": [
          {
            "name": "WashCompanyId",
            "in": "path",
            "description": "WashCompanyId for get washers list",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/integerId"
            }
          },
          {
            "name": "searchName",
            "in": "query",
            "required": true,
            "description": "Washer's name",
            "schema": {
              "type": "string",
              "properties": {
                "dateFrom": {
                  "type": "string",
                  "required": true
                }
              }
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": true,
            "description": "Pagination page number (start from 1)",
            "schema": {
              "type": "integer",
              "properties": {
                "page": {
                  "type": "integer",
                  "default": 1
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authenticated"
          },
          "400": {
            "description": "Validation error"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/washer": {
      "get": {
        "tags": [
          "Washers"
        ],
        "summary": "Get washer info by id",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "WasherId for get info",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/integerId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authenticated"
          },
          "400": {
            "description": "Validation error"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/{washerId}/getOrders": {
      "get": {
        "tags": [
          "Washers"
        ],
        "summary": "Get washer orders",
        "parameters": [
          {
            "name": "washerId",
            "in": "path",
            "description": "WasherId for get orders",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/integerId"
            }
          },
          {
            "name": "isActive",
            "in": "query",
            "required": true,
            "description": "Orders status",
            "schema": {
              "type": "boolean",
              "properties": {
                "IsActive": {
                  "type": "boolean",
                  "default": true
                }
              }
            }
          },
          {
            "name": "dateFrom",
            "in": "query",
            "required": true,
            "description": "Orders from date (year-month-day)",
            "schema": {
              "type": "string",
              "properties": {
                "dateFrom": {
                  "type": "string",
                  "default": "year-month-day",
                  "required": true
                }
              }
            }
          },
          {
            "name": "dateTo",
            "in": "query",
            "required": true,
            "description": "Orders to date (year-month-day)",
            "schema": {
              "type": "string",
              "properties": {
                "dateTo": {
                  "type": "string",
                  "default": "year-month-day",
                  "required": true
                }
              }
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": true,
            "description": "Pagination page number (start from 1)",
            "schema": {
              "type": "integer",
              "properties": {
                "page": {
                  "type": "integer",
                  "default": 1
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authenticated"
          },
          "400": {
            "description": "Validation error"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/{WashCompanyId}/insertWasher": {
      "post": {
        "tags": [
          "Washers"
        ],
        "summary": "Add new washer to company",
        "parameters": [
          {
            "name": "WashCompanyId",
            "in": "path",
            "description": "CompanyId for adding new washer",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/integerId"
            }
          }
        ],
        "requestBody": {
          "description": "Add new washer for this company",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/addWasher"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authenticated"
          },
          "400": {
            "description": "Validation error"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/{washerId}/updateWasher": {
      "post": {
        "tags": [
          "Washers"
        ],
        "summary": "Edit data about washer",
        "parameters": [
          {
            "name": "washerId",
            "in": "path",
            "description": "WasherId for get orders",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/integerId"
            }
          }
        ],
        "requestBody": {
          "description": "Edit washer",
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/editWasher"
              }
            }
          }

        },
        "responses": {
          "200": {
            "description": "Authenticated"
          },
          "400": {
            "description": "Validation error"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    }
  }
}